magento/magento2#10045 Rename item resolver to item provider
