+++ app/code/Magento/Sitemap/Model/ItemProvider/ConfigReaderInterface.php
@@ -4,7 +4,7 @@
  * See COPYING.txt for license details.
  */
 
-namespace Magento\Sitemap\Model\ItemResolver;
+namespace Magento\Sitemap\Model\ItemProvider;
 
 /**
  * Item resolver config reader interface
