+++ app/code/Magento/Sitemap/Model/ItemProvider/ProductConfigReader.php
@@ -4,18 +4,18 @@
  * See COPYING.txt for license details.
  */
 
-namespace Magento\Sitemap\Model\ItemResolver;
+namespace Magento\Sitemap\Model\ItemProvider;
 
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Store\Model\ScopeInterface;
 
-class CategoryConfigReader implements ConfigReaderInterface
+class ProductConfigReader implements ConfigReaderInterface
 {
     /**#@+
      * Xpath config settings
      */
-    const XML_PATH_CHANGE_FREQUENCY = 'sitemap/category/changefreq';
-    const XML_PATH_PRIORITY = 'sitemap/category/priority';
+    const XML_PATH_CHANGE_FREQUENCY = 'sitemap/product/changefreq';
+    const XML_PATH_PRIORITY = 'sitemap/product/priority';
     /**#@-*/
 
     /**
