+++ app/code/Magento/Sitemap/Model/Sitemap.php
@@ -14,7 +14,7 @@ use Magento\Framework\DataObject;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\UrlInterface;
 use Magento\Robots\Model\Config\Value;
-use Magento\Sitemap\Model\ItemResolver\ItemResolverInterface;
+use Magento\Sitemap\Model\ItemProvider\ItemProviderInterface;
 use Magento\Sitemap\Model\ResourceModel\Sitemap as SitemapResource;
 
 /**
@@ -164,9 +164,9 @@ class Sitemap extends \Magento\Framework\Model\AbstractModel implements \Magento
     /**
      * Item resolver
      *
-     * @var ItemResolverInterface
+     * @var ItemProviderInterface
      */
-    private $itemResolver;
+    private $itemProvider;
 
     /**
      * Sitemap config reader
@@ -201,7 +201,7 @@ class Sitemap extends \Magento\Framework\Model\AbstractModel implements \Magento
      * @param \Magento\Framework\Data\Collection\AbstractDb|null $resourceCollection
      * @param array $data
      * @param DocumentRoot|null $documentRoot
-     * @param ItemResolverInterface|null $itemResolver
+     * @param ItemProviderInterface|null $itemProvider
      * @param SitemapConfigReaderInterface|null $configReader
      * @param \Magento\Sitemap\Model\SitemapItemInterfaceFactory|null $sitemapItemFactory
      * @SuppressWarnings(PHPMD.ExcessiveParameterList)
@@ -223,7 +223,7 @@ class Sitemap extends \Magento\Framework\Model\AbstractModel implements \Magento
         \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
         array $data = [],
         \Magento\Config\Model\Config\Reader\Source\Deployed\DocumentRoot $documentRoot = null,
-        ItemResolverInterface $itemResolver = null,
+        ItemProviderInterface $itemProvider = null,
         SitemapConfigReaderInterface $configReader = null,
         \Magento\Sitemap\Model\SitemapItemInterfaceFactory $sitemapItemFactory = null
     ) {
@@ -238,7 +238,7 @@ class Sitemap extends \Magento\Framework\Model\AbstractModel implements \Magento
         $this->_storeManager = $storeManager;
         $this->_request = $request;
         $this->dateTime = $dateTime;
-        $this->itemResolver = $itemResolver ?: ObjectManager::getInstance()->get(ItemResolverInterface::class);
+        $this->itemProvider = $itemProvider ?: ObjectManager::getInstance()->get(ItemProviderInterface::class);
         $this->configReader = $configReader ?: ObjectManager::getInstance()->get(SitemapConfigReaderInterface::class);
         $this->sitemapItemFactory = $sitemapItemFactory ?: ObjectManager::getInstance()->get(
             \Magento\Sitemap\Model\SitemapItemInterfaceFactory::class
@@ -278,7 +278,7 @@ class Sitemap extends \Magento\Framework\Model\AbstractModel implements \Magento
      * @param DataObject $sitemapItem
      * @return $this
      * @deprecated
-     * @see ItemResolverInterface
+     * @see ItemProviderInterface
      */
     public function addSitemapItem(DataObject $sitemapItem)
     {
@@ -292,7 +292,7 @@ class Sitemap extends \Magento\Framework\Model\AbstractModel implements \Magento
      *
      * @return void
      * @deprecated
-     * @see ItemResolverInterface
+     * @see ItemProviderInterface
      */
     public function collectSitemapItems()
     {
@@ -332,7 +332,7 @@ class Sitemap extends \Magento\Framework\Model\AbstractModel implements \Magento
      */
     protected function _initSitemapItems()
     {
-        $sitemapItems = $this->itemResolver->getItems($this->getStoreId());
+        $sitemapItems = $this->itemProvider->getItems($this->getStoreId());
         $mappedItems = $this->mapToSitemapItem();
         $this->_sitemapItems = array_merge($sitemapItems, $mappedItems);
