+++ app/code/Magento/Sitemap/etc/di.xml
@@ -7,7 +7,7 @@
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <preference for="Magento\Sitemap\Model\SitemapItemInterface" type="Magento\Sitemap\Model\SitemapItem" />
-    <preference for="Magento\Sitemap\Model\ItemResolver\ItemResolverInterface" type="Magento\Sitemap\Model\ItemResolver\Composite" />
+    <preference for="Magento\Sitemap\Model\ItemProvider\ItemProviderInterface" type="Magento\Sitemap\Model\ItemProvider\Composite" />
     <preference for="Magento\Sitemap\Model\SitemapConfigReaderInterface" type="Magento\Sitemap\Model\SitemapConfigReader" />
 
     <type name="Magento\Sitemap\Model\Sitemap">
@@ -22,28 +22,28 @@
             </argument>
         </arguments>
     </type>
-    <type name="Magento\Sitemap\Model\ItemResolver\Composite">
+    <type name="Magento\Sitemap\Model\ItemProvider\Composite">
         <arguments>
-            <argument name="itemResolvers" xsi:type="array">
-                <item name="categoryResolver" xsi:type="object">Magento\Sitemap\Model\ItemResolver\Category</item>
-                <item name="cmsPageResolver" xsi:type="object">Magento\Sitemap\Model\ItemResolver\CmsPage</item>
-                <item name="productResolver" xsi:type="object">Magento\Sitemap\Model\ItemResolver\Product</item>
+            <argument name="itemProviders" xsi:type="array">
+                <item name="categoryProvider" xsi:type="object">Magento\Sitemap\Model\ItemProvider\Category</item>
+                <item name="cmsPageProvider" xsi:type="object">Magento\Sitemap\Model\ItemProvider\CmsPage</item>
+                <item name="productProvider" xsi:type="object">Magento\Sitemap\Model\ItemProvider\Product</item>
             </argument>
         </arguments>
     </type>
-    <type name="Magento\Sitemap\Model\ItemResolver\Category">
+    <type name="Magento\Sitemap\Model\ItemProvider\Category">
         <arguments>
-            <argument name="configReader" xsi:type="object">Magento\Sitemap\Model\ItemResolver\CategoryConfigReader</argument>
+            <argument name="configReader" xsi:type="object">Magento\Sitemap\Model\ItemProvider\CategoryConfigReader</argument>
         </arguments>
     </type>
-    <type name="Magento\Sitemap\Model\ItemResolver\Product">
+    <type name="Magento\Sitemap\Model\ItemProvider\Product">
         <arguments>
-            <argument name="configReader" xsi:type="object">Magento\Sitemap\Model\ItemResolver\ProductConfigReader</argument>
+            <argument name="configReader" xsi:type="object">Magento\Sitemap\Model\ItemProvider\ProductConfigReader</argument>
         </arguments>
     </type>
-    <type name="Magento\Sitemap\Model\ItemResolver\CmsPage">
+    <type name="Magento\Sitemap\Model\ItemProvider\CmsPage">
         <arguments>
-            <argument name="configReader" xsi:type="object">Magento\Sitemap\Model\ItemResolver\CmsPageConfigReader</argument>
+            <argument name="configReader" xsi:type="object">Magento\Sitemap\Model\ItemProvider\CmsPageConfigReader</argument>
         </arguments>
     </type>
 </config>
